+++ tools/webTools/lib/src/initWebPage.js
@@ -59,7 +59,7 @@ window[MODULE].push(`
            <label calss="card-title">SVN地址：</label>
            <p class="card-text">
              <div class="form-group">
-              <input type="email" class="form-control" aria-describedby="emailHelp" placeholder="输入svn地址..">
+              <input class="form-control" placeholder="输入svn地址..">
               <small id="emailHelp" class="form-text text-muted">SVN地址为portal日后扫描文件的地方，请正确设置</small>
              </div>
              <a href="javascript:void(0);" onClick="javascript:setSVNAddress(this);" class="btn btn-outline-primary btn-sm pull-right">提交</a>