+++ History.md
@@ -330,3 +330,17 @@ commit by afterloe (lm6289511@gmail.com)
 > tools/webTools/lib/src/initWebPage.js  
 > tools/webTools/lib/src/util.js  
 
+
+1.0.8 / 2017-3-6 15:32:26
+==================
+commit by afterloe (lm6289511@gmail.com)
+
+  * master: 完成单节点设置svn地址
+
+
+### modify files
+> services/workflowService.js  
+> tools/webTools/dataManager.html  
+> tools/webTools/lib/src/initWebPage.js  
+> tools/webTools/lib/src/util.js  
+